Got Date creation working
