+++ app/src/main/AndroidManifest.xml
@@ -16,8 +16,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".MoodSelected"
-            android:parentActivityName=".MainActivity"></activity>
     </application>
 
 </manifest>
