+++ app/src/main/java/cs/ualberta/ca/neagu_feelsbook/CurrentMood.java
@@ -15,10 +15,10 @@ public class CurrentMood implements MoodInterface{
         this.message = "Default Message";
     }
 
-    CurrentMood(Date date){
+    CurrentMood(Date date, String message){
         this.mood = "Sad Mood ";
         this.moodDate = date;
-        this.message = "Default Message";
+        this.message = message;
     }
 
     public String getMood() {
@@ -51,16 +51,16 @@ public class CurrentMood implements MoodInterface{
         }
     }
 
-    // Need to override toString class because otherwise this defaults to
-    // the items toString class which just return the id of the element as a string
-    @Override
-    public String toString(){
-        String returnString = new String();
-
-        returnString = this.getMood();
-        returnString += this.getMoodDate();
-
-        return returnString;
-    }
+//    // Need to override toString class because otherwise this defaults to
+//    // the items toString class which just return the id of the element as a string
+//    @Override
+//    public String toString(){
+//        String returnString = new String();
+//
+//        returnString = this.getMood();
+//        returnString += this.getMoodDate();
+//
+//        return returnString;
+//    }
 
 }
